Refactor the resolver classes into a package and add one for dnspython. Fixes #580
