+++ CHANGES.rst
@@ -32,6 +32,9 @@
 - Travis CI tests on Python 3.7.0a4 and PyPy 2.7 5.10.0 and PyPy 3.5
   5.10.1.
 
+- Add the ``dnspython`` resolver as a lightweight alternative to
+  c-ares.
+
 1.3a1 (2018-01-27)
 ==================
