+++ dev-requirements.txt
@@ -15,6 +15,8 @@ coveralls>=1.0
 # See version requirements in setup.py
 cffi
 futures
+dnspython
+idna
 # Makes tests faster
 psutil
 # For viewing README.rst (restview --long-description),
