+++ doc/dns.rst
@@ -22,19 +22,21 @@ A resolver implements the 5 standandard functions from the
 Configuration
 =============
 
-gevent includes three implementations of resolvers, and applications
+gevent includes four implementations of resolvers, and applications
 can provide their own implementation. By default, gevent uses
-:class:`gevent.resolver_thread.Resolver`.
+:class:`gevent.resolver.thread.Resolver`.
 
 Configuration can be done through the ``GEVENT_RESOLVER`` environment
-variable. Specify ``ares``, ``thread``, or ``block`` to use the
-:class:`gevent.resolver_ares.Resolver`,
-:class:`gevent.resolver_thread.Resolver`, or
-:class:`gevent.socket.BlockingResolver`, respectively, or set it to
+variable. Specify ``ares``, ``thread``, ``dnspython``, or ``block`` to use the
+:class:`gevent.resolver.ares.Resolver`,
+:class:`gevent.resolver.thread.Resolver`, or
+:class:`gevent.resolver.dnspython.Resolver`, or
+:class:`gevent.resolver.blocking.Resolver`, respectively, or set it to
 the fully-qualified name of an implementation of the standard
 functions.
 
 .. toctree::
 
-   gevent.resolver_thread
-   gevent.resolver_ares
+   gevent.resolver.thread
+   gevent.resolver.ares
+   gevent.resolver.dnspython
