+++ src/greentest/greentest/sysinfo.py
@@ -126,3 +126,6 @@ CONN_ABORTED_ERRORS.append(ECONNRESET)
 CONN_ABORTED_ERRORS = frozenset(CONN_ABORTED_ERRORS)
 
 RESOLVER_ARES = os.getenv('GEVENT_RESOLVER') == 'ares'
+RESOLVER_DNSPYTHON = os.getenv('GEVENT_RESOLVER') == 'dnspython'
+
+RESOLVER_NOT_SYSTEM = RESOLVER_ARES or RESOLVER_DNSPYTHON
