+++ src/greentest/tests_that_dont_monkeypatch.txt
@@ -1,0 +1,15 @@
+test___example_servers.py
+test__backdoor.py
+test__example_echoserver.py
+test__example_udp_client.py
+test__getaddrinfo_import.py
+test__example_portforwarder.py
+test__pywsgi.py
+test__server.py
+test__server_pywsgi.py
+test__socket_close.py
+test__socket_dns6.py
+test__socket_errors.py
+test__socket_send_memoryview.py
+test__socket_timeout.py
+test__examples.py
